% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{allrefs}
\alias{allrefs}
\title{Get all the datasets from each of the six celltypes in the ABRIG data.}
\usage{
allrefs(con, pfiles, genelocs, gene)
}
\arguments{
\item{con}{A duckdb connection}

\item{genelocs}{}

\item{gene}{A gene  from the human genome.}
}
\value{
A list of data table by type for the given gene.
}
\description{
Get all the datasets from each of the six celltypes in the ABRIG data.
}
\examples{
pfiles <<- ABRIGparquet_paths()
utils::data("gloc_hg19", package = "GGIpack")
allrefs(gene = 'DSP', pfiles = pfiles, genelocs = gloc_hg19)
}
