% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{checkData}
\alias{checkData}
\title{Takes in a path to a data table and coherses the data into the proper format}
\usage{
checkData(path)
}
\arguments{
\item{path}{The complete path to the DN8 file.}
}
\description{
Takes in a path to a data table and coherses the data into the proper format
}
