% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ensg}
\alias{ensg}
\title{A instance of Ensembl gene for human genome 19}
\format{
data object
}
\usage{
ensg
}
\description{
A instance of Ensembl gene for human genome 19
}
\keyword{datasets}
