% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{filterByRange}
\alias{filterByRange}
\title{filter a GGI (ABRIGresource) instance by range}
\usage{
filterByRange(res, ggr, tag, radius = 1e+05, ggr_field = "gene_name")
}
\arguments{
\item{res}{ABRIGresource instance}

\item{ggr}{GenomicRanges instance}

\item{tag}{character(1) value in `ggr_field` used for filtering, e.g., a gene}

\item{radius}{numeric(1) flanking region size}

\item{ggr_field}{character(1) metadata element in mcols(ggr) for filtering}
}
\description{
filter a GGI (ABRIGresource) instance by range
}
