% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ops.R
\name{makeGWASTrack}
\alias{makeGWASTrack}
\title{make GWASTrack for shiny app.}
\usage{
makeGWASTrack(name = "NA", dat)
}
\arguments{
\item{name}{the name of the track in the shiny app}

\item{dat}{the data table exported from the find_data}
}
\value{
a gwasTrack read to graph by igvshiny
}
\description{
make GWASTrack for shiny app.
}
\note{
It is required that the column  header need to be the following:
c("SNP", "CHR", "BP", "A1", "A2", "gene", "geneId", "statistic", 
 "P", "FDR", "BETA", "SE", "MAF")
 It is  advised if the desired data does not have these column headers that the column headers need to be changed to the above.
}
\examples{
dat = find_data()
makeGWASTrack(dat = dat)
}
