% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpleIGV.R
\name{simpleIGV}
\alias{simpleIGV}
\title{app with a button to add specific GWAS track}
\usage{
simpleIGV()
}
\description{
app with a button to add specific GWAS track
}
\examples{
if (interactive()) {
  simpleIGV()
}
}
