% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tinyapp.R
\name{tinyapp}
\alias{tinyapp}
\title{demo app}
\usage{
tinyapp(con, genelocs = genes(EnsDb.Hsapiens.v75))
}
\arguments{
\item{con}{a DBI connection}

\item{genelocs}{a GRanges instance with gene addresses}
}
\description{
demo app
}
